Text/button same time and cleaned up
I made it so you can change the color of the text by holding down and tap to edit the text. I also changed constraints so that everything is fixed. The keyboard will not mess anything up now. Also added how to add to scrollview at top in comment form.